+++ history.md
@@ -1,6 +1,7 @@
 ## Edge
 
 * Add new command `wti rmlocale locale_code_1 locale_code_2 ...` to delete a locale from a project.
+* Add new command `wti rm path_to_file1 path_to_file2 ...` to delete a master file from a project.
 
 ## Version 1.8.3 / 2011-11-14